Deploy swipe tabs for landing page, refactor join room activity
