+++ app/src/main/java/hk/ust/cse/hunkim/questionroom/CreateRoomFragment.java
@@ -1,0 +1,21 @@
+package hk.ust.cse.hunkim.questionroom;
+
+import android.os.Bundle;
+import android.support.v4.app.Fragment;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+/**
+ * Created by CAI on 17/11/2015.
+ */
+public class CreateRoomFragment extends Fragment {
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+
+        View rootView = inflater.inflate(R.layout.fragment_create_room, container, false);
+
+        return rootView;
+    }
+}
