+++ app/src/main/res/layout/activity_join.xml
@@ -1,79 +1,5 @@
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
+<android.support.v4.view.ViewPager xmlns:android="http://schemas.android.com/apk/res/android"
+    android:id="@+id/pager"
     android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:gravity="center_horizontal"
-    android:orientation="vertical"
-    android:paddingBottom="@dimen/activity_vertical_margin"
-    android:paddingLeft="@dimen/activity_horizontal_margin"
-    android:paddingRight="@dimen/activity_horizontal_margin"
-    android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="hk.ust.cse.hunkim.questionroom.JoinActivity"
-    android:weightSum="1">
-
-    <AutoCompleteTextView
-        android:id="@+id/room_name"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:hint="@string/prompt_room_name"
-        android:inputType="textShortMessage"
-        android:maxLines="1"
-        android:singleLine="true"
-        style="@style/EditTextCustomHolo"/>
-
-    <Button
-        android:id="@+id/join_button"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="16dp"
-        android:onClick="attemptJoin"
-        android:text="@string/action_join"
-        android:textStyle="bold"
-        style="@style/ButtonCustomHolo"/>
-
-    <Button
-        android:id="@+id/create_room"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="16dp"
-        android:onClick="attemptJoin"
-        android:text="@string/create_room"
-        android:textStyle="bold"
-        style="@style/ButtonCustomHolo" />
-
-    <Button
-        android:id="@+id/all_rooms"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="16dp"
-        android:text="All Rooms"
-        android:textStyle="bold"
-        style="@style/ButtonCustomHolo" />
-
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="30dp"
-        android:textAppearance="?android:attr/textAppearanceMedium"
-        android:text="@string/question_room"
-        android:id="@+id/textView"
-        android:layout_gravity="center_horizontal"
-        android:textSize="30dp" />
-
-    <TextView
-        android:layout_width="298dp"
-        android:layout_height="wrap_content"
-        android:textAppearance="?android:attr/textAppearanceSmall"
-        android:text="@string/question_room_descibe"
-        android:id="@+id/question_room_descibe"
-        android:textStyle="bold"
-        android:layout_marginTop="30dp"
-        android:layout_gravity="center_horizontal"
-        android:textSize="20dp"
-        android:singleLine="false"
-        android:password="false"
-        android:editable="false"
-        android:gravity="center" />
-
-</LinearLayout>
+    android:layout_height="match_parent" />
